Find gaps in correspondence between products
Add a script `gaps.py` to the `digitalearthau/scripts` folder. It
provides a CLI interface to the function `find_gaps` and dumps
the generated summary in `yaml` to `stdout` (or optionally a file).
A 'gap' here arises when a dataset in one of the products does not
have a corresponding dataset having the same spatio-temporal domain
in the other product.
The user should specify products to compare by name, and may
also specify a date range for the datasets.