Simplify tag handling
The TaskApp handling stuff in DEA was creating a task_datetime and using
that for the work directory, but other bits of code were expecting a tag
to be passed in on every invocation.

For now, lets go with the TaskApp task_datetime, and use that to generate
a task_tag.