Remove git submodules for including dea-notebooks
The next commit will include an alternative way of pulling down the
documentation from the `dea-notebooks` repo.

git submodules point to a specific commit in a remote repo, which is
usually what you want. However, we always want to point to the latest
commit in `dea-notebooks`, since the worst that happens is the
documentation doesn't update until a breakage is fixed.