Make Travis install ODC through pypi
If we want to test it against dev versions of ODC, we
could create a separate optional runner for that.

Currently, installing through the package repo breaks the
build here too often, and everyone's too busy to fix it. The
breakages haven't been due to ODC bugs, only due to packaging
oddities.