Update all redirects to point to DEA Docs Homepage
This was originally the documentation home page for digital earth
australia, hosted at geoscienceaustralia.github.io. We moved it to
docs.dea.ga.gov.au back in 2019 or so and I changed all the HTML files
here to redirect to their new location.

However, I made a mistake and files with spaces in them weren't handled correctly.

As well, since then, the structure of the documentation has changed and
95% of the redirects broke.

Google is currently finding one of the broken redirect pages as the
first result for "dc.find_datasets".

This should automatically redirect someone who does end up at one of
these pages, as well as let google know that all pages have permanently
moved.